+++ objc2/src/declare.rs
@@ -19,7 +19,7 @@
 //! use objc2::foundation::NSObject;
 //! use objc2::rc::{Id, Owned};
 //! use objc2::runtime::{Class, Object, Sel};
-//! use objc2::{class, sel, msg_send, msg_send_id, ClassType};
+//! use objc2::{sel, msg_send, msg_send_id, ClassType};
 //! # #[cfg(feature = "gnustep-1-7")]
 //! # unsafe { objc2::__gnustep_hack::get_class_to_force_linkage() };
 //!